Separate pollAndUserService to own services
